writecoach/requirements.txt
streamlit
fastapi
uvicorn
pydantic
nltk
openai
anthropic
google-generativeai
pandas
plotly
python-dotenv